requirements.txt
streamlit==1.28.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.7
plotly==5.16.1
pandas==2.0.3

# Enhanced UI components
streamlit-option-menu==0.3.6
streamlit-lottie==0.0.5
streamlit-elements==0.1.0
streamlit-ace==0.1.1
streamlit-chat==0.1.1
streamlit-aggrid==0.3.4
streamlit-autorefresh==0.0.6
extra-streamlit-components==0.1.60

# Image processing
Pillow==10.0.0

# Data science and ML
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.2
matplotlib==3.7.2
seaborn==0.12.2

# Real-time features
asyncio-mqtt==0.11.1
websockets==11.0.3
redis==4.6.0
schedule==1.2.0
watchdog==3.0.0

# Security and performance
cryptography==41.0.4
pyjwt==2.8.0
psutil==5.9.5
bcrypt==4.0.1

# Database
sqlalchemy==2.0.20
alembic==1.12.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
cachetools==5.3.1
